#3 for aditional fuctions

+++ GL/Scene_2.cpp
@@ -22,7 +22,7 @@ namespace Scene2_Variables {
 
 	std::random_device rd;
 	std::mt19937 gen(rd());
-	std::uniform_real_distribution<double> dis(0.0, 1.0);
+	std::uniform_real_distribution<float> dis(0.0, 1.0);
 
 	float Rect1_ratio = 1.0f;
 	float Rect2_ratio = 1.0f;
